% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/cheerio.R
\name{cheerio}
\alias{cheerio}
\title{Parse HTML/XML}
\usage{
cheerio(html, ...)
}
\arguments{
\item{html}{Character vector html/xml}
}
\description{
This function parses a HTML or XML document in a private V8 context.
}

